Test poll delay
